Don't try to close files that don't exist.

+++ app/jobs/bg_s3uploadable/uploader_job.rb
@@ -32,8 +32,10 @@ module BgS3uploadable
         record.save validate: false
       end
     ensure
-      f.close
-      f.unlink
+      if f
+        f.close
+        f.unlink
+      end
     end
   end
 end
